set up a feign client in the ticket service with the price.  with posting a new ticket, grab the request param id to look up a ticket price, which will get back a price object from the price microservice.  and then sets the price in the controller.

+++ ticketservice/src/main/java/com/udemy/bharath/ticketservice/Ticket.java
@@ -5,6 +5,7 @@ import lombok.*;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 @Entity
 @NoArgsConstructor @AllArgsConstructor @ToString
@@ -15,7 +16,9 @@ public class Ticket {
 	private Long id;
 	@NotNull
 	private String event;
-	@NotNull private Long price;
-	@NotNull private String status;
+	@NotNull
+	private BigDecimal price;
+	@NotNull
+	private String status;
 }
